refactor(sponsors): rename the useSponsorData hook to useGetSponsorData
